Add dependency to load values in dropdown controls in right order
The dropdown controls created by. the ComponentsList "processRelationships" depends on the Variables Lists. When loading a JASP file, the dropdown controls must get their values after the Variables lists have got their own values. This dependency is not detected because the Dropdown controls are created dynamically (only when the ComponentsList is created). So we have to set the dependencies explicitly.